Refactor describe opts into their own file
