+++ internal/plugin/plugin_manager.go
@@ -144,75 +144,6 @@ func (p *PluginManager) Collect(ctx context.Context, partition *config.Partition
 	return CollectResponseFromProto(collectResponse), nil
 }
 
-// formatToProto takes a config.Format, describes the format and returns the proto.FormatData for the plugin
-func (p *PluginManager) formatToProto(ctx context.Context, format *config.Format, tablePlugin *pplugin.Plugin) (*proto.FormatData, error) {
-	//  check if the format is provided by the table plugin or whether we need to start format plugin
-	formatPluginName, ok := config.GetPluginForFormat(format)
-	if !ok {
-		return nil, fmt.Errorf("error determining source plugin for format %s", format.FullName)
-	}
-	// if the format is provided by the table plugin, we can just convert it to proto
-	if formatPluginName == tablePlugin.Plugin {
-		slog.Info("format is provided by the table plugin - converting to proto", "format", format.FullName, "plugin", formatPluginName)
-		return format.ToProto(), nil
-	}
-
-	// so the plugin is NOT the table plugin - start it if needed
-	formatPlugin := pplugin.NewPlugin(formatPluginName)
-	if formatPlugin.Plugin == tablePlugin.Plugin {
-		// the format is provided by the table plugin - nothing to do
-		return nil, nil
-	}
-	slog.Info("format is not provided by the table plugin - describing the format and converting to a regex format", "format", format.FullName, "plugin", formatPlugin.Plugin)
-
-	// so the format is provided by a different plugin - start it if needed and execute a describe
-	// if format is NOT a preset, we need to pass the custom formats to the describe call
-	var opts []DescribeOpts
-	if format.PresetName == "" {
-		opts = append(opts, WithCustomFormats(format), WithCustomFormatsOnly())
-	}
-
-	describeResponse, err := p.Describe(ctx, formatPlugin.Plugin, opts...)
-	if err != nil {
-		return nil, fmt.Errorf("error resolving format: %w", err)
-	}
-
-	var desc *types.FormatDescription
-	if format.PresetName != "" {
-		// if format is a preset
-		desc, ok = describeResponse.FormatPresets[format.PresetName]
-		if !ok {
-			return nil, fmt.Errorf("plugin '%s' returned no description for format preset '%s'", formatPluginName, format.PresetName)
-		}
-	} else {
-		// we expect the custom format to be the one we asked for
-		desc, ok = describeResponse.CustomFormats[format.FullName]
-		if !ok {
-			return nil, fmt.Errorf("plugin '%s' returned no description for format '%s'", formatPluginName, format.PresetName)
-		}
-	}
-
-	return &proto.FormatData{Name: format.FullName, Regex: desc.Regex}, nil
-}
-
-type DescribeOpts func(*proto.DescribeRequest)
-
-func WithCustomFormats(customFormats ...*config.Format) DescribeOpts {
-	return func(req *proto.DescribeRequest) {
-		var customFormatsProto []*proto.FormatData
-		for _, f := range customFormats {
-			customFormatsProto = append(customFormatsProto, f.ToProto())
-		}
-		req.CustomFormats = customFormatsProto
-	}
-}
-
-func WithCustomFormatsOnly() DescribeOpts {
-	return func(req *proto.DescribeRequest) {
-		req.CustomFormatsOnly = true
-	}
-}
-
 // Describe starts the plugin if needed, and returns the plugin description, including description of any custom formats
 func (p *PluginManager) Describe(ctx context.Context, pluginName string, opts ...DescribeOpts) (*types.DescribeResponse, error) {
 	// build plugin ref from the name
@@ -278,6 +209,57 @@ func (p *PluginManager) UpdateCollectionState(ctx context.Context, partition *co
 
 	// just return - the observer is responsible for waiting for completion
 	return err
+}
+
+// formatToProto takes a config.Format, describes the format and returns the proto.FormatData for the plugin
+func (p *PluginManager) formatToProto(ctx context.Context, format *config.Format, tablePlugin *pplugin.Plugin) (*proto.FormatData, error) {
+	//  check if the format is provided by the table plugin or whether we need to start format plugin
+	formatPluginName, ok := config.GetPluginForFormat(format)
+	if !ok {
+		return nil, fmt.Errorf("error determining source plugin for format %s", format.FullName)
+	}
+	// if the format is provided by the table plugin, we can just convert it to proto
+	if formatPluginName == tablePlugin.Plugin {
+		slog.Info("format is provided by the table plugin - converting to proto", "format", format.FullName, "plugin", formatPluginName)
+		return format.ToProto(), nil
+	}
+
+	// so the plugin is NOT the table plugin - start it if needed
+	formatPlugin := pplugin.NewPlugin(formatPluginName)
+	if formatPlugin.Plugin == tablePlugin.Plugin {
+		// the format is provided by the table plugin - nothing to do
+		return nil, nil
+	}
+	slog.Info("format is not provided by the table plugin - describing the format and converting to a regex format", "format", format.FullName, "plugin", formatPlugin.Plugin)
+
+	// so the format is provided by a different plugin - start it if needed and execute a describe
+	// if format is NOT a preset, we need to pass the custom formats to the describe call
+	var opts []DescribeOpts
+	if format.PresetName == "" {
+		opts = append(opts, WithCustomFormats(format), WithCustomFormatsOnly())
+	}
+
+	describeResponse, err := p.Describe(ctx, formatPlugin.Plugin, opts...)
+	if err != nil {
+		return nil, fmt.Errorf("error resolving format: %w", err)
+	}
+
+	var desc *types.FormatDescription
+	if format.PresetName != "" {
+		// if format is a preset
+		desc, ok = describeResponse.FormatPresets[format.PresetName]
+		if !ok {
+			return nil, fmt.Errorf("plugin '%s' returned no description for format preset '%s'", formatPluginName, format.PresetName)
+		}
+	} else {
+		// we expect the custom format to be the one we asked for
+		desc, ok = describeResponse.CustomFormats[format.FullName]
+		if !ok {
+			return nil, fmt.Errorf("plugin '%s' returned no description for format '%s'", formatPluginName, format.PresetName)
+		}
+	}
+
+	return &proto.FormatData{Name: format.FullName, Regex: desc.Regex}, nil
 }
 
 func (p *PluginManager) getSourcePluginReattach(ctx context.Context, partition *config.Partition, tablePlugin *pplugin.Plugin) (*proto.SourcePluginReattach, error) {
